Hurt FX

+++ Assets/Source/DamageSystem/DamageInfo.cs
@@ -4,7 +4,7 @@ namespace Quinn.DamageSystem
 {
 	public record DamageInfo
 	{
-		public int Damage = 1;
+		public float Damage = 1f;
 		public Vector2 Direction;
 		public TeamType TeamType;
 	}
